Build with optimizations
Build it more like what VitaSDK does, optimizing for the Vita's CPU. `-O1` is the maximum optimization that doesn't crash, for some reason. The other parameters are the ones used by VitaSDK for their build of GCC.

This is a small piece of code, but removing the `-g` parameter and bringing it more in line with the rest of the things built can be benefitial.